* develop:
  fix: 설정 수정 - queryDsl 설정 수정 - application-prod.yml 설정값 추가
  refactor: 줄바꿈 수정
  url 주소 수정
  test : 미팅 결과에 따른 코드 리팩토링에 의한 테스트 코드 수정
  refactor : 미팅 결과에 따른 코드 리팩토링
  feat: CORS 설정 추가
  fix: redirectUrl 수정
  docs: Swagger 상의 Contact 삭제
  fix: 소셜 로그인, 로그아웃 시 redirectUrl 주소 변경
  test : 마이프로필 수정 및 회원 탈퇴 테스트 코드 작성
  feat : 마이프로필 수정 및 회원 탈퇴 기능 구현

+++ src/main/java/com/zerototen/savegame/service/RecordService.java
@@ -1,15 +1,18 @@
 package com.zerototen.savegame.service;
 
 import com.zerototen.savegame.domain.dto.CreateRecordServiceDto;
+import com.zerototen.savegame.domain.dto.RecordAnalysisServiceDto;
 import com.zerototen.savegame.domain.dto.UpdateRecordServiceDto;
+import com.zerototen.savegame.domain.dto.response.RecordAnalysisResponse;
 import com.zerototen.savegame.domain.dto.response.RecordResponse;
 import com.zerototen.savegame.domain.dto.response.ResponseDto;
+import com.zerototen.savegame.domain.entity.Member;
 import com.zerototen.savegame.domain.entity.Record;
-import com.zerototen.savegame.exception.CustomException;
 import com.zerototen.savegame.exception.ErrorCode;
 import com.zerototen.savegame.repository.MemberRepository;
 import com.zerototen.savegame.repository.RecordRepository;
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import javax.transaction.Transactional;
@@ -25,19 +28,23 @@ public class RecordService {
     private final MemberRepository memberRepository;
     private final RecordRepository recordRepository;
 
+    // 지출 내역 등록
     @Transactional
     public ResponseDto<?> create(CreateRecordServiceDto serviceDto) {
-        memberRepository.findById(serviceDto.getMemberId())
-            .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_MEMBER));
+        Member member = memberRepository.findById(serviceDto.getMemberId()).orElse(null);
+        if (member == null) {
+            return ResponseDto.fail(ErrorCode.NOT_FOUND_MEMBER.getDetail());
+        }
 
         log.debug("Create record -> memberId: {}", serviceDto.getMemberId());
         return ResponseDto.success(recordRepository.save(serviceDto.toEntity()));
     }
 
+    // 지출 내역 조회 (가계부 메인)
     public ResponseDto<?> getInfos(Long memberId, LocalDate startDate, LocalDate endDate,
         List<String> categories) {
         if (startDate.isAfter(endDate)) {
-            throw new CustomException(ErrorCode.STARTDATE_AFTER_ENDDATE);
+            return ResponseDto.fail(ErrorCode.STARTDATE_AFTER_ENDDATE.getDetail());
         }
         List<Record> records = recordRepository.findByMemberIdAndUseDateDescWithOptional(
             memberId, startDate, endDate, categories);
@@ -45,30 +52,39 @@ public class RecordService {
         return ResponseDto.success(records.stream().map(RecordResponse::from).collect(Collectors.toList()));
     }
 
+    // 지출 내역 분석 (가계부 분석)
     public ResponseDto<?> getAnalysisInfo(Long memberId, int year, int month) {
         LocalDate startDate = LocalDate.of(year, month, 1);
         LocalDate endDate = startDate.withDayOfMonth(startDate.lengthOfMonth());
 
-        return ResponseDto.success(
-            recordRepository.findByMemberIdAndUseDateAndAmountSumDesc(memberId, startDate, endDate)
-                .stream().map(i -> {
-                    if (i.getCategory() == null) {
-                        throw new CustomException(ErrorCode.CATEGORY_IS_NULL);
-                    }
-                    if (i.getTotal() == null || i.getTotal() <= 0) {
-                        throw new CustomException(ErrorCode.INVALID_TOTAL);
-                    }
-                    return i.toResponse();
-                }).collect(Collectors.toList()));
+        List<RecordAnalysisServiceDto> serviceDtos = recordRepository.findByMemberIdAndUseDateAndAmountSumDesc(memberId,
+            startDate, endDate);
+
+        List<RecordAnalysisResponse> responses = new ArrayList<>();
+
+        for (RecordAnalysisServiceDto serviceDto : serviceDtos) {
+            if (serviceDto.getCategory() == null) {
+                return ResponseDto.fail(ErrorCode.CATEGORY_IS_NULL.getDetail());
+            }
+            if (serviceDto.getTotal() == null || serviceDto.getTotal() <= 0) {
+                return ResponseDto.fail(ErrorCode.INVALID_TOTAL.getDetail());
+            }
+            responses.add(serviceDto.toResponse());
+        }
+
+        return ResponseDto.success(responses);
     }
 
+    // 지출 내역 수정
     @Transactional
     public ResponseDto<?> update(UpdateRecordServiceDto serviceDto) {
-        Record record = recordRepository.findById(serviceDto.getId())
-            .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_RECORD));
+        Record record = recordRepository.findById(serviceDto.getId()).orElse(null);
+        if (record == null) {
+            return ResponseDto.fail(ErrorCode.NOT_FOUND_RECORD.getDetail());
+        }
 
         if (!record.getMemberId().equals(serviceDto.getMemberId())) {
-            throw new CustomException(ErrorCode.NOT_MATCH_MEMBER);
+            return ResponseDto.fail(ErrorCode.NOT_MATCH_MEMBER.getDetail());
         }
 
         record.update(serviceDto);
@@ -76,13 +92,16 @@ public class RecordService {
         return ResponseDto.success("Update Success");
     }
 
+    // 지출 내역 삭제
     @Transactional
     public ResponseDto<?> delete(Long id, Long memberId) {
-        Record record = recordRepository.findById(id)
-            .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_RECORD));
+        Record record = recordRepository.findById(id).orElse(null);
+        if (record == null) {
+            return ResponseDto.fail(ErrorCode.NOT_FOUND_RECORD.getDetail());
+        }
 
         if (!record.getMemberId().equals(memberId)) {
-            throw new CustomException(ErrorCode.NOT_MATCH_MEMBER);
+            return ResponseDto.fail(ErrorCode.NOT_MATCH_MEMBER.getDetail());
         }
 
         recordRepository.delete(record);
